Moved NodeNorm/NameRes into own files.
